Move accessor_base.hpp into accessor/detail
Refactor because it is not related to buffer.
